requirements.txt
google-generativeai
beautifulsoup4
requests
dotenv